+++ clb-app/backend/app/routers/budget.py
@@ -36,7 +36,7 @@ def get_daily_budget(request: Request) -> BudgetStatus:
     # Add standalone session costs for today
     session_costs = sum(c["amount"] for c in request.app.state.daily_session_costs if c["date"] == today_str)
     
-    total = float(events_total + session_costs)
+    total = int(events_total + session_costs)
     is_overdrafted, overdraft_amount, remaining = detect_overdraft(total, DAILY_BUDGET)
     
     # Weekly total also needs updated logic
@@ -77,19 +77,19 @@ def get_weekly_budget(request: Request) -> dict:
         
     # Initialize all 7 days of the week for the heatmap
     week_dates = [week_start + timedelta(days=i) for i in range(7)]
-    totals = {d.strftime("%Y-%m-%d"): 0.0 for d in week_dates}
+    totals = {d.strftime("%Y-%m-%d"): 0 for d in week_dates}
     
     # Add event costs (preferring actual_cost from Sage Mode)
     for event in events:
         day_key = event.start_time.strftime("%Y-%m-%d")
         if day_key in totals:
             cost = (event.actual_cost if event.actual_cost is not None else event.calculated_cost) or 0
-            totals[day_key] += float(cost)
+            totals[day_key] += int(cost)
             
     # Add standalone session costs
     for session in request.app.state.daily_session_costs:
         if session["date"] in totals:
-            totals[session["date"]] += float(session["amount"])
+            totals[session["date"]] += int(session["amount"])
     weekly_total = sum(totals.values())
     return {
         "daily_totals": totals,
